move deploy lib to its' own folder
